Refactor `Alert` model
