+++ functions/src/models/index.ts
@@ -3,7 +3,7 @@ import Event from './Event';
 import Participant from './Participant';
 import Phone from './Phone';
 import Geo from './Geo';
-export type { DBAlert } from './Alert';
+export type { DBAlert } from './Alert/types';
 export type { DBEvent } from './Event';
 export type { ParticipantArgs } from './Participant';
 export type { VerificationStatus } from './Phone';
